Add clicked URL and clicked email hooks
To make the website URL and author email address
clickable in the about dialogue, we need to install
callbacks defining the actions to be performed.
What I have here is incredibly hackish and should
not be used as an example, but should work for
people with GNOME and/or xdg-utils installed.

Signed-off-by: Philip Allison <sane@not.co.uk>