Speed up board rendering by only drawing occupied squares
Mark the drawing area for the board as "app paintable" (basically a
hint to the library that we've overridden the expose event handler).
Render the empty board to a static pixmap and set it as the widget's
background, so we don't have to redraw it every time the window is
exposed - just when it's resized.

Turn off Cairo's antialiasing - we want nicely distinguishable
squares, please.

Signed-off-by: Philip Allison <sane@not.co.uk>