Finish implementing hexagonal board support in BoardState
We now have proper initialisation for hexagonal boards, and a new value
in the "piece" enumerated type for non-existent squares (holes). Added
range checking and a function for calculating the adjacency of surrounding
squares (for valid move checking & piece capturing) mean that Game and
GameBoard needn't care about the shape of the board itself.

Signed-off-by: Philip Allison <sane@not.co.uk>