Allow user to choose between 2 player hexagonal or 2/4 player square games
Add a "board shape" drop-down to the new game dialogue, and a property
accessor for it which is used to set the game type accordingly, instead
of hard-coding to hexagonal.  Also, move setting the default state of the
new game dialogue into the NewGameDialog class, for cleanliness.
Enforce valid game type constraints on the new game dialogue, so the user
can't - for example - set up a four-player game on a hexagonal board.

Signed-off-by: Philip Allison <sane@not.co.uk>