Start work on hexagonal board rendering, plus general rendering improvements
Render the backaground of hexagonal boards correctly (but not yet the
pieces or valid move highlights).  Also, convert the square board rendering
to draw one big square in one of the two colours, then overlay squares of
the other, rather than draw many individual squares.  This now looks nicer
with antialiasing enabled than the latter, so re-enable it. :)

Signed-off-by: Philip Allison <sane@not.co.uk>