Keep track of scores, and show status notifications during & after game
Game object keeps track of each player's score.  GameWindow now listens
for the move_made signal, and uses score information to display current
game status in the status bar.  The status bar is now actually a row of
several bars, as this is the simplest and visually cleanest way to have
multiple messages side-by-side.

Signed-off-by: Philip Allison <sane@not.co.uk>