Back to the drawing board - standard minimax AI
Caching future moves was too memory intensive and error prone, so scrap
the whole scheme and go back to standard minimax.  Still seems to have
a bug or two to iron out; for example, the computer won't make a move
to eliminate another player if it can, and doesn't play particularly
cleverly in a four-player game.  Its opening moves are also very
predictable.

Signed-off-by: Philip Allison <sane@not.co.uk>