Move piece capture logic into BoardState and flesh out two-player AI
Some game logic has been moved to BoardState so that AI can be
implemented with less hassle.  Standard minimax AI has now been
implemented, and works well for the two player game, but not so
well for the four-player game.  It takes too long without some form
of pruning, allowing us only to look one move ahead, and has a weak
evaluation function - this means a player cannot be truly risk averse;
they may capture a lot of pieces, but don't seem to spot the risk of
any position they may be put in.  Either that or it's just downright
bug-ridden...

Signed-off-by: Philip Allison <sane@not.co.uk>