Add some useful debug output to the AI
Can see what's going on now.  In the two-player game, the "suicidal"
green player *is* actually doing the best it can in the end game
(it may as well go out with 3 pieces captured than with 2 grown),
so the questions that remain are:

1) Why so slow?
2) Why does it play such a crap opening game?

Number 1 will hopefully be solved by more sensible code in BoardState
and alpha-beta pruning.  Number 2 is most likely the ridiculously naive
position evaluation function.

Signed-off-by: Philip Allison <sane@not.co.uk>