Go back to the drawing board and implement a simpler, faster AI concept
AI is no longer based on minimax, instead it works purely by looking at
the current player's possible moves, concentrating on a detailed board
state evaluation function instead of simulating opponent's moves.  The
end result is a lot faster than minimax because of the dramatic reduction
in number of moves evaluated, and seems to play the game on the square
board reasonably well, but for some reason isn't too brilliant on the
hexagonal game board.

Also fix a couple of bugs so that the game functions correctly when player
1 is an AI player, and tweak the move signalling code of AI players so
that it's easier to see what moves are being made.

Signed-off-by: Philip Allison <sane@not.co.uk>