Send game details out to clients as they change
Wrap client IOChannels inside a ClientSocket object, which handles
non-blocking writes in the background on behalf of calling code, but
which makes a const reference to the IOChannel accessible so that
calling code can connect its own event handlers as well if it wishes.
Reimplement client handling in ServerStatusDialog to use ClientSockets,
and create a method for serialising game details into a binary blob,
sending it out to all connected clients whenever details change.

Signed-off-by: Philip Allison <sane@not.co.uk>