Make game board always sensitive to clicks
To simplify various bits of code, the game board should *always* be
sensitive to user input; the test for whether or not to act on it is in
onClick itself.  This fixes the inability to make moves when the first
player in a game isn't a local human (i.e. they're a network client or
AI player).

Signed-off-by: Philip Allison <philip.allison@smoothwall.net>