MinGW tweaks; bump to 0.2
Make the application compile and link cleanly
under MinGW.  This includes bumping the required
versions of autoconf/automake, as we need (for
example) a newer aclocal than was previously
being asked for if we're going to successfully
rebuild the build scripts.  Tentatively bump
the version number to 0.2 to reflect recent work.

Signed-off-by: Philip Allison <sane@not.co.uk>