Changes to data files to accomodate Windows
On Windows, we want pre-rendered icon files and
about dialogue images.  Also, the "stock_people"
image seems to be missing from the GTK+ runtime
package for Windows, so include it here.  Change
the data makefile to not bother doing the bits
which aren't necessary on Windows, such as
rendering out various icon sizes and installing
a .desktop file.

Signed-off-by: Philip Allison <sane@not.co.uk>