Use Socket pointers; put local client info in bold
Make serversock a simple pointer to a Socket, not a Glib::RefPtr, to
bring in line with the interface now provided by ServerStatusDialog.
Also, show information pertaining to the local client in bold in the
game details display, so that players can distinguish which colour they
are currently assigned to.

Signed-off-by: Philip Allison <sane@not.co.uk>