Make start & end pop-ups trigger on main loop idle
Prevent sending the last move of a game over the network from being
delayed until the win dialogue is closed by the winning player, by
connecting a function for displaying the dialogue to Glib::SignalIdle.
Use the same technique for the "You are playing as..." dialogues on
clients at the start of a network game.

Signed-off-by: Philip Allison <sane@not.co.uk>