Fix end-game logic bug reported by Jan Engelhardt
When the game ends, the winning player should automatically capture
any empty squares, as this advances the board to the state it would be
in if the game was played until the board was full.  Without this, it
is possible for one player to block off all their opponent's moves and
lose the game, even though the number of empty squares remaining would
make their piece count highest if captured.

Signed-off-by: Philip Allison <sane@not.co.uk>